lib: remove serializer-factory
