+++ lib/record-serialization-fallback.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const serializerFactory = require('./serializer-factory');
+const serialize = require('./json5-serialize');
 
 const jsrs = {};
 module.exports = jsrs;
@@ -8,7 +8,7 @@ module.exports = jsrs;
 // Serialize a JavaScript value using the JSTP Record Serialization format
 // and return a string representing it.
 //
-jsrs.stringify = serializerFactory.createSerializer();
+jsrs.stringify = serialize;
 
 // Deserialize a string in the JSTP Record Serialization format into
 // a JavaScript value and return it.
