+++ lib/record-serialization.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const safeRequire = require('./common').safeRequire;
-const serializerFactory = require('./serializer-factory');
+const serialize = require('./json5-serialize');
 
 const jsrs = {};
 module.exports = jsrs;
@@ -19,7 +19,7 @@ const jstpNative =
 if (jstpNative) {
   Object.assign(jsrs, jstpNative);
   if (!USE_NATIVE_SERIALIZER) {
-    jsrs.stringify = serializerFactory.createSerializer();
+    jsrs.stringify = serialize;
   }
 } else {
   console.warn(
